# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "请求状态码定义\ncode= 0, 调用成功\ncode=-1, 系统错误\ncode= 1, 提示接口返回的message\ncode=10, 登录token过期\ncode=20, 接口权限错误,没有权限访问该接口",
        "title": "go-template(项目名)",
        "termsOfService": "http://localhost:8000/swagger/index.html",
        "contact": {},
        "license": {
            "name": "Apache 2.0"
        },
        "version": "v1.2"
    },
    "host": "localhost:8000",
    "paths": {
        "/api/demo/getUserById": {
            "get": {
                "description": "get user info by id",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Business-Group"
                ],
                "summary": "get user by id,返参为结构体对象",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "用户ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/models.HTTPError"
                        }
                    }
                }
            }
        },
        "/api/demo/hello": {
            "get": {
                "description": "say hello to given name",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "demo样例"
                ],
                "summary": "say hello,返参data为基本类型",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户名",
                        "name": "name",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/models.Result"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/models.HTTPError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.HTTPError": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "响应编码",
                    "type": "integer",
                    "format": "int",
                    "default": -1
                },
                "message": {
                    "description": "消息",
                    "type": "string",
                    "format": "string",
                    "default": "请求失败"
                },
                "success": {
                    "description": "是否请求成功",
                    "type": "boolean",
                    "format": "bool",
                    "default": false
                }
            }
        },
        "models.Result": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "响应编码",
                    "type": "integer",
                    "format": "int"
                },
                "data": {
                    "description": "响应结果",
                    "format": "object"
                },
                "message": {
                    "description": "消息",
                    "type": "string",
                    "format": "string",
                    "default": "ok"
                },
                "success": {
                    "description": "是否请求成功",
                    "type": "boolean",
                    "format": "bool"
                },
                "time": {
                    "description": "请求时间",
                    "type": "string",
                    "format": "string"
                }
            }
        },
        "models.User": {
            "type": "object",
            "properties": {
                "age": {
                    "description": "年龄",
                    "type": "integer",
                    "default": 27
                },
                "email": {
                    "description": "邮箱",
                    "type": "string",
                    "default": "JoeOK@gmail.com"
                },
                "id": {
                    "description": "ID",
                    "type": "integer",
                    "format": "int64",
                    "default": 1
                },
                "name": {
                    "description": "姓名",
                    "type": "string",
                    "default": "Joe"
                },
                "nation": {
                    "description": "国籍",
                    "type": "string",
                    "default": "USA"
                },
                "title": {
                    "description": "职位",
                    "type": "string",
                    "default": "engineer"
                }
            }
        }
    }
}